moved new functions to composer module, renamed
