+++ plonk-core/src/constraint_system/boolean.rs
@@ -8,7 +8,6 @@
 
 use crate::constraint_system::{StandardComposer, Variable};
 use ark_ec::{PairingEngine, TEModelParameters};
-use ark_ff::Field;
 use num_traits::{One, Zero};
 
 impl<E, P> StandardComposer<E, P>
@@ -49,52 +48,6 @@ where
 
         a
     }
-
-    /// A gate which outputs a variable whose value is 1 if
-    /// the input is 0 and whose value is 0 otherwise
-    pub fn is_zero_gate(&mut self, a: Variable) -> Variable {
-        // Get relevant field values
-        let a_value = self.variables.get(&a).unwrap();
-        let y_value = a_value.inverse().unwrap_or_else(E::Fr::one);
-
-        // This has value 1 if input value is zero, value 0 otherwise
-        let b_value = E::Fr::one() - *a_value * y_value;
-
-        let y = self.add_input(y_value);
-        let b = self.add_input(b_value);
-        let zero = self.zero_var();
-
-        // Enforce constraints. The constraint system being used here is
-        // a * y + b - 1 = 0
-        // a * b = 0
-        // where y is auxiliary and b is the boolean (a == 0).
-        let _a_times_b = self.arithmetic_gate(|gate| {
-            gate.witness(a, b, Some(zero)).mul(E::Fr::one())
-        });
-
-        let _first_constraint = self.arithmetic_gate(|gate| {
-            gate.witness(a, y, Some(zero))
-                .mul(E::Fr::one())
-                .fan_in_3(E::Fr::one(), b)
-                .constant(-E::Fr::one())
-        });
-
-        b
-    }
-
-    /// A gate which outputs a variable whose value is 1 if the
-    /// two input variables have equal values and whose value is 0 otherwise.
-    pub fn is_eq_gate(&mut self, a: Variable, b: Variable) -> Variable {
-        let difference = self.arithmetic_gate(|gate| {
-            gate.witness(a, b, None).add(E::Fr::one(), -E::Fr::one())
-        });
-        // TODO: These dummy constraints are a hack to avoid a 2^k issue,
-        // specifically the q_arith poly being constant.
-        // It should be safe to remove this dummy constraint once that issue
-        // is resolved.
-        self.add_dummy_constraints();
-        self.is_zero_gate(difference)
-    }
 }
 
 #[cfg(test)]
@@ -105,68 +58,6 @@ mod test {
     use ark_bls12_377::Bls12_377;
     use ark_bls12_381::Bls12_381;
     use num_traits::One;
-
-    fn test_correct_is_zero_gate<E, P>()
-    where
-        E: PairingEngine,
-        P: TEModelParameters<BaseField = E::Fr>,
-    {
-        // Check that it gives true on zero input:
-        let res = gadget_tester(
-            |composer: &mut StandardComposer<E, P>| {
-                let one = composer.add_input(E::Fr::one());
-                let is_zero = composer.is_zero_gate(composer.zero_var());
-                composer.assert_equal(is_zero, one);
-            },
-            32,
-        );
-
-        // Check that it gives false on non-zero input:
-        let res2 = gadget_tester(
-            |composer: &mut StandardComposer<E, P>| {
-                let one = composer.add_input(E::Fr::one());
-                let is_zero = composer.is_zero_gate(one);
-                composer.assert_equal(is_zero, composer.zero_var());
-            },
-            32,
-        );
-
-        assert!(res.is_ok() && res2.is_ok())
-    }
-
-    fn test_correct_is_eq_gate<E, P>()
-    where
-        E: PairingEngine,
-        P: TEModelParameters<BaseField = E::Fr>,
-    {
-        // Check that it gives true on equal inputs:
-        let res = gadget_tester(
-            |composer: &mut StandardComposer<E, P>| {
-                let one = composer.add_input(E::Fr::one());
-
-                let field_element = E::Fr::one().double();
-                let a = composer.add_input(field_element);
-                let b = composer.add_input(field_element);
-                let is_eq = composer.is_eq_gate(a, b);
-                composer.assert_equal(is_eq, one);
-            },
-            32,
-        );
-
-        // Check that it gives false on non-equal inputs:
-        let res2 = gadget_tester(
-            |composer: &mut StandardComposer<E, P>| {
-                let field_element = E::Fr::one().double();
-                let a = composer.add_input(field_element);
-                let b = composer.add_input(field_element.double());
-                let is_eq = composer.is_eq_gate(a, b);
-                composer.assert_equal(is_eq, composer.zero_var());
-            },
-            32,
-        );
-
-        assert!(res.is_ok() && res2.is_ok())
-    }
 
     fn test_correct_bool_gate<E, P>()
     where
@@ -205,8 +96,6 @@ mod test {
     // Test for Bls12_381
     batch_test!(
         [
-            test_correct_is_zero_gate,
-            test_correct_is_eq_gate,
             test_correct_bool_gate,
             test_incorrect_bool_gate
         ],
@@ -219,8 +108,6 @@ mod test {
     // Test for Bls12_377
     batch_test!(
         [
-            test_correct_is_zero_gate,
-            test_correct_is_eq_gate,
             test_correct_bool_gate,
             test_incorrect_bool_gate
         ],
