+++ plonk-core/src/constraint_system/composer.rs
@@ -19,6 +19,7 @@ use crate::permutation::Permutation;
 use alloc::collections::BTreeMap;
 use ark_ec::models::TEModelParameters;
 use ark_ec::PairingEngine;
+use ark_ff::Field;
 #[cfg(feature = "trace")]
 use ark_ff::{BigInteger, PrimeField};
 use core::marker::PhantomData;
@@ -336,6 +337,47 @@ where
             E::Fr::zero(),
             None,
         );
+    }
+
+    /// A gate which outputs a variable whose value is 1 if
+    /// the input is 0 and whose value is 0 otherwise
+    pub fn is_zero_with_output(&mut self, a: Variable) -> Variable {
+        // Get relevant field values
+        let a_value = self.variables.get(&a).unwrap();
+        let y_value = a_value.inverse().unwrap_or_else(E::Fr::one);
+
+        // This has value 1 if input value is zero, value 0 otherwise
+        let b_value = E::Fr::one() - *a_value * y_value;
+
+        let y = self.add_input(y_value);
+        let b = self.add_input(b_value);
+        let zero = self.zero_var();
+
+        // Enforce constraints. The constraint system being used here is
+        // a * y + b - 1 = 0
+        // a * b = 0
+        // where y is auxiliary and b is the boolean (a == 0).
+        let _a_times_b = self.arithmetic_gate(|gate| {
+            gate.witness(a, b, Some(zero)).mul(E::Fr::one())
+        });
+
+        let _first_constraint = self.arithmetic_gate(|gate| {
+            gate.witness(a, y, Some(zero))
+                .mul(E::Fr::one())
+                .fan_in_3(E::Fr::one(), b)
+                .constant(-E::Fr::one())
+        });
+
+        b
+    }
+
+    /// A gate which outputs a variable whose value is 1 if the
+    /// two input variables have equal values and whose value is 0 otherwise.
+    pub fn is_eq_with_output(&mut self, a: Variable, b: Variable) -> Variable {
+        let difference = self.arithmetic_gate(|gate| {
+            gate.witness(a, b, None).add(E::Fr::one(), -E::Fr::one())
+        });
+        self.is_zero_with_output(difference)
     }
 
     /// Conditionally selects a [`Variable`] based on an input bit.
@@ -704,6 +746,68 @@ mod test {
         assert!(res.is_ok());
     }
 
+    fn test_correct_is_zero_with_output<E, P>()
+    where
+        E: PairingEngine,
+        P: TEModelParameters<BaseField = E::Fr>,
+    {
+        // Check that it gives true on zero input:
+        let res = gadget_tester(
+            |composer: &mut StandardComposer<E, P>| {
+                let one = composer.add_input(E::Fr::one());
+                let is_zero = composer.is_zero_with_output(composer.zero_var());
+                composer.assert_equal(is_zero, one);
+            },
+            32,
+        );
+
+        // Check that it gives false on non-zero input:
+        let res2 = gadget_tester(
+            |composer: &mut StandardComposer<E, P>| {
+                let one = composer.add_input(E::Fr::one());
+                let is_zero = composer.is_zero_with_output(one);
+                composer.assert_equal(is_zero, composer.zero_var());
+            },
+            32,
+        );
+
+        assert!(res.is_ok() && res2.is_ok())
+    }
+
+    fn test_correct_is_eq_with_output<E, P>()
+    where
+        E: PairingEngine,
+        P: TEModelParameters<BaseField = E::Fr>,
+    {
+        // Check that it gives true on equal inputs:
+        let res = gadget_tester(
+            |composer: &mut StandardComposer<E, P>| {
+                let one = composer.add_input(E::Fr::one());
+
+                let field_element = E::Fr::one().double();
+                let a = composer.add_input(field_element);
+                let b = composer.add_input(field_element);
+                let is_eq = composer.is_eq_with_output(a, b);
+                composer.assert_equal(is_eq, one);
+            },
+            32,
+        );
+
+        // Check that it gives false on non-equal inputs:
+        let res2 = gadget_tester(
+            |composer: &mut StandardComposer<E, P>| {
+                let field_element = E::Fr::one().double();
+                let a = composer.add_input(field_element);
+                let b = composer.add_input(field_element.double());
+                let is_eq = composer.is_eq_with_output(a, b);
+                composer.assert_equal(is_eq, composer.zero_var());
+            },
+            32,
+        );
+
+        assert!(res.is_ok() && res2.is_ok())
+    }
+
     fn test_conditional_select<E, P>()
     where
         E: PairingEngine,
@@ -821,6 +925,8 @@ mod test {
         [
             test_initial_circuit_size,
             test_prove_verify,
+            test_correct_is_zero_with_output,
+            test_correct_is_eq_with_output,
             test_conditional_select,
             test_multiple_proofs
         ],
@@ -835,6 +941,8 @@ mod test {
         [
             test_initial_circuit_size,
             test_prove_verify,
+            test_correct_is_zero_with_output,
+            test_correct_is_eq_with_output,
             test_conditional_select,
             test_multiple_proofs
         ],
